text selection finished
text selected on a website is adeqautely validated and the end and start word of the selection is autofilled. The autofill must follow specific rules for navigating the structure of the selection object that is created. This means that it will have to function in multiple different ways to achieve the same effect.